backend/requirements.txt
pymongo[srv]
motor
fastapi[standard]
uvicorn[standard]
pydantic
pydantic[email]
python-dotenv
httpx
pyjwt
pwdlib[argon2]